Move Data logic to Context
